requirements.txt
selenium==4.10.0
python-telegram-bot==13.7
python-dotenv==0.19.0